pyproject.toml
[tool.poetry]
name = "codecollector"
version = "0.1.0"
description = "Collect code from your dir to one .txt file to send it to GPT"
authors = ["werserk <werserk@inbox.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
